backend/requirements/base.txt
Django==5.0.3
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
django-tenants==3.6.1
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
python-decouple==3.8
httpx==0.27.0
Pillow==10.2.0
dj-database-url==2.1.0
openpyxl==3.1.2
xhtml2pdf==0.2.13
qrcode[pil]==7.4.2
